requirements.txt
streamlit==1.44.1
llama-index==0.12.28
llama-index-core==0.12.28
llama-index-embeddings-huggingface==0.5.2
llama-index-llms-groq==0.3.1
pandas==2.2.3
python-dotenv==1.1.0
sentence-transformers==4.0.2  
groq==0.22.0
Flask
gunicorn
